# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WishShop API",
        "contact": {
            "email": "lijy91@foxmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "WishShop Server (Local)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "WishShop Server (Local)"
        },
        {
            "url": "https://wishshop-api-dev.thecode.me",
            "description": "WishShop Server (Development)"
        },
        {
            "url": "https://wishshop-api.thecode.me",
            "description": "WishShop Server (Production)"
        }
    ],
    "paths": {
        "/account/register": {
            "post": {
                "tags": [
                    "account"
                ],
                "operationId": "App\\Http\\Controllers\\AccountController::register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/account/login": {
            "post": {
                "tags": [
                    "account"
                ],
                "operationId": "App\\Http\\Controllers\\AccountController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/account/logout": {
            "post": {
                "tags": [
                    "account"
                ],
                "operationId": "App\\Http\\Controllers\\AccountController::logout",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "jwt_auth": []
                    }
                ]
            }
        },
        "/admin/stores/{store_id}/categories": {
            "get": {
                "tags": [
                    "admin/categories"
                ],
                "summary": "Get all Categories",
                "operationId": "App\\Http\\Controllers\\Admin\\CategoryController::index",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Parent Id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/admin/stores/{store_id}/categories/{category_id}": {
            "get": {
                "tags": [
                    "admin/categories"
                ],
                "summary": "Get a specific category",
                "operationId": "App\\Http\\Controllers\\Admin\\CategoryController::show",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "Category Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/admin/stores/{store_id}/products": {
            "get": {
                "tags": [
                    "admin/products"
                ],
                "summary": "Get all Products",
                "operationId": "App\\Http\\Controllers\\Admin\\ProductController::index",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/admin/stores/{store_id}/products/{product_id}": {
            "get": {
                "tags": [
                    "admin/products"
                ],
                "summary": "Get a specific product",
                "operationId": "App\\Http\\Controllers\\Admin\\ProductController::show",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/admin/users/{user_id}": {
            "get": {
                "tags": [
                    "admin/users"
                ],
                "summary": "Get a specific user",
                "operationId": "App\\Http\\Controllers\\Admin\\UserController::show",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "jwt_auth": {
                "type": "apiKey",
                "description": "To make authenticated requests via http using the built in methods, you will need to set an authorization header as follows:\\n```\\nAuthorization: Bearer {yourtokenhere}\\n```\\n",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}